Fixed view compile issue

+++ app/compilers/Blade/Engine.php
@@ -19,7 +19,7 @@ class BladeEngine extends Compiler
 		$this->blade = new BladeCompiler();
 		$this->contents = $this->blade->compile($contents);
 
-		if ($this->contents != file_get_contents(APP . 'storage/cache' . DS . $view . '.php'))
+		if (file_exists(APP . 'storage/cache' . DS . $view . '.php') && $this->contents != file_get_contents(APP . 'storage/cache' . DS . $view . '.php'))
 		{
 			file_put_contents(APP . 'storage/cache' . DS . $view . '.php', $this->contents);
 		}
